Rescope FImGuiInputProcessor to FImGuiContext
Creating an InputProcessor per overlay is unnecessary and can result in one not being created if a context is created before an SImGuiOverlay is created.

This also resolves a bug where all ImGui windows for a given context would have SetKeyboardFocus() called for them, since the ActiveWindow was not being taken into account when performing selection. See the bottom of `FImGuiInputProcessor::Tick` for the new implementation.
